backend/apps/finance/views.py
from rest_framework import viewsets
from .models import Payment, Expense
from .serializers import PaymentSerializer, ExpenseSerializer

class PaymentViewSet(viewsets.ModelViewSet):
    queryset = Payment.objects.all()
    serializer_class = PaymentSerializer

class ExpenseViewSet(viewsets.ModelViewSet):
    queryset = Expense.objects.all()
    serializer_class = ExpenseSerializer
